feat: implementa endpoints de livro

+++ src/services/autor.service.js
@@ -1,5 +1,9 @@
 // Repositories
 import AutorRepository from "../repositories/autor.repository.js";
+import LivroRepository from "../repositories/livro.repository.js";
+
+// Utils
+import { errorHandler } from "../utils/error-handler.js";
 
 async function createAutor(autor) {
   return await AutorRepository.insertAutor(autor);
@@ -14,15 +18,14 @@ async function getAutor(autorId) {
 }
 
 async function deleteAutor(autorId) {
-  // TODO: Descomentar após implementar LivroRepository
-  // const livros = await LivroRepository.listLivrosByAutorId(autorId);
-  // if (livros.length > 0) {
-  // const error = errorHandler(
-  // 405,
-  // `Não é possível excluir um autor que possui um ou mais livros cadastrados.`
-  // );
-  // throw error;
-  // }
+  const livros = await LivroRepository.listLivrosByAutorId(autorId);
+  if (livros.length > 0) {
+    const error = errorHandler(
+      405,
+      `Não é possível excluir um autor que possui um ou mais livros cadastrados.`
+    );
+    throw error;
+  }
 
   await AutorRepository.deleteAutor(autorId);
 }
